feat(libraries/modules): move commands.ts
